+++ pkg/deployment/reconcile/plan_builder_test.go
@@ -79,6 +79,8 @@ type testContext struct {
 
 	Inspector inspectorInterface.Inspector
 	state     member.StateInspector
+
+	AgencyState state.State
 }
 
 func (c *testContext) IsFeatureImageSupported(feature features.Feature) bool {
@@ -183,7 +185,7 @@ func (c *testContext) WithAgencyCache(action func(state.State)) bool {
 }
 
 func (c *testContext) GetAgencyCache() (state.State, bool) {
-	return state.State{}, true
+	return c.AgencyState, true
 }
 
 func (c *testContext) SecretsModInterface() generic.ModClient[*core.Secret] {